# docs/swagger.yaml
definitions:
...
  github_com_evrone_go-clean-template_internal_entity.AdminBulkDeleteRequest:
    properties:
      userIds:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - userIds
    type: object
  github_com_evrone_go-clean-template_internal_entity.AdminBulkDeleteResponse:
    properties:
      deleted:
        type: integer
    type: object
  github_com_evrone_go-clean-template_internal_entity.AdminBulkStatusRequest:
    properties:
      status:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminUserStatus'
      userIds:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - status
    - userIds
    type: object
  github_com_evrone_go-clean-template_internal_entity.AdminBulkStatusResponse:
    properties:
      updated:
        type: integer
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.AdminInviteRequest:
    properties:
      email:
        type: string
      message:
        type: string
      role:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminUserRole'
    required:
    - email
    - role
    type: object
  github_com_evrone_go-clean-template_internal_entity.AdminInviteResponse:
    properties:
      expiresAt:
        type: string
      invited:
        type: boolean
    type: object
  github_com_evrone_go-clean-template_internal_entity.AdminLoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  github_com_evrone_go-clean-template_internal_entity.AdminProfile:
    properties:
      createdAt:
        type: string
      displayName:
        type: string
      email:
        type: string
      id:
        type: string
      permissions:
        items:
          type: string
        type: array
      primaryExam:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ExamCategory'
      role:
        type: string
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.AdminUser:
    properties:
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      phoneNumber:
        type: string
      role:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminUserRole'
      status:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminUserStatus'
      updatedAt:
        type: string
      username:
        type: string
    type: object
  github_com_evrone_go-clean-template_internal_entity.AdminUserCreateRequest:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
      phoneNumber:
        type: string
      role:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminUserRole'
      status:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminUserStatus'
      username:
        type: string
    required:
    - email
    - firstName
    - lastName
    - password
    - phoneNumber
    - role
    - status
    - username
    type: object
  github_com_evrone_go-clean-template_internal_entity.AdminUserList:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminUser'
        type: array
      meta:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminUsersMeta'
    type: object
  github_com_evrone_go-clean-template_internal_entity.AdminUserRole:
    enum:
    - superadmin
    - admin
    - manager
    - cashier
    type: string
    x-enum-varnames:
    - AdminUserRoleSuperAdmin
    - AdminUserRoleAdmin
    - AdminUserRoleManager
    - AdminUserRoleCashier
  github_com_evrone_go-clean-template_internal_entity.AdminUserStatus:
    enum:
    - active
    - inactive
    - invited
    - suspended
    type: string
    x-enum-varnames:
    - AdminUserStatusActive
    - AdminUserStatusInactive
    - AdminUserStatusInvited
    - AdminUserStatusSuspended
  github_com_evrone_go-clean-template_internal_entity.AdminUserUpdateRequest:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
      phoneNumber:
        type: string
      role:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminUserRole'
      status:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminUserStatus'
      username:
        type: string
    type: object
  github_com_evrone_go-clean-template_internal_entity.AdminUsersMeta:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.AuthResponse:
    properties:
      accessToken:
        type: string
      user:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.User'
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.ExamCategory:
    enum:
    - NEET_PG
    - NEET_UG
    - JEE
    - UPSC
    type: string
    x-enum-varnames:
    - ExamCategoryNEETPG
    - ExamCategoryNEETUG
    - ExamCategoryJEE
    - ExamCategoryUPSC
...
  github_com_evrone_go-clean-template_internal_entity.ExamProfile:
    properties:
      currentStreakDays:
        type: integer
      exam:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ExamCategory'
      lastLoginAt:
        type: string
      longestStreakDays:
        type: integer
      overallLevel:
        type: integer
      totalCorrect:
        type: integer
      totalQuestions:
        type: integer
      totalTimeSeconds:
        type: integer
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.MeResponse:
    properties:
      examProfile:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ExamProfile'
      user:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.User'
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.TelegramAuthRequest:
    properties:
      displayName:
        type: string
      exam:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ExamCategory'
      telegramId:
        type: string
    required:
    - exam
    - telegramId
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.User:
    properties:
      createdAt:
        type: string
      displayName:
        type: string
      email:
        type: string
      id:
        type: string
      primaryExam:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ExamCategory'
      role:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.UserRole'
      telegramId:
        type: string
    type: object
  github_com_evrone_go-clean-template_internal_entity.UserRole:
    enum:
    - USER
    - ADMIN
    - SUPER_ADMIN
    type: string
    x-enum-varnames:
    - UserRoleUser
    - UserRoleAdmin
    - UserRoleSuperAdmin
...
  internal_controller_http_v1.ErrorResponse:
    properties:
      error:
        example: message
        type: string
    type: object
info:
  contact: {}
paths:
...
  /admin/users:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Comma separated statuses
        in: query
        name: status
        type: string
      - description: Role
        in: query
        name: role
        type: string
      - description: Username search
        in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminUserList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: List admin users
      tags:
      - 'Admin: Users'
    post:
      consumes:
      - application/json
      parameters:
      - description: User payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminUserCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Create admin user
      tags:
      - 'Admin: Users'
  /admin/users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Delete admin user
      tags:
      - 'Admin: Users'
    patch:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminUserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Update admin user
      tags:
      - 'Admin: Users'
  /admin/users/bulk-delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bulk delete payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminBulkDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminBulkDeleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Bulk delete users
      tags:
      - 'Admin: Users'
  /admin/users/bulk-status:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bulk status payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminBulkStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminBulkStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Bulk status update
      tags:
      - 'Admin: Users'
  /admin/users/invite:
    post:
      consumes:
      - application/json
      parameters:
      - description: Invite payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminInviteRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminInviteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Invite admin user
      tags:
      - 'Admin: Users'
  /auth/admin/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Admin credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      summary: Login as admin
      tags:
      - Auth
  /auth/admin/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminProfile'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Admin profile
      tags:
      - Auth
  /auth/telegram:
    post:
      consumes:
      - application/json
      parameters:
      - description: Telegram payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.TelegramAuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      summary: Login or signup via Telegram
      tags:
      - Auth
...
  /me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.MeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - UserAuth: []
      summary: Get current user and profile
      tags:
      - 'App: User'
...
swagger: "2.0"